third_party/WebKit/LayoutTests/http/tests/loading/dont-preload-non-img-srcset.html

<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
    window.targetScaleFactor = 1;
</script>
<script src="../resources/srcset-helper.js"></script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
</head>
<body>
<script>
    shouldBeTrue("internals.isPreloaded('resources/base-image1.png');");
    shouldBeFalse("internals.isPreloaded('resources/empty.html');");
    shouldBeFalse("internals.isPreloaded('resources/empty.js');");
</script>
<div>This test passes if the image is preloaded, but none of the non-img elements with a srcset attribute is.</div>
<img height="100" width="100" src="resources/preload-test.jpg" srcset="resources/base-image1.png 1x, resources/base-image3.png 3x, resources/base-image2.png 2x"></img>
<iframe srcset="resources/empty.html"></iframe>
<script srcset="resources/empty.js"></script>
</body>
</html>
